requirements.txt
Django==2.1
django-extensions==2.1.2
django-simple-authentication==0.1.2
django-simple-pages==0.2.1
django-watchman==0.15.0
psycopg2==2.7.3.2
requests==2.19.1